generic ServiceInterface with implementation for firecracker
