+++ vhive.go
@@ -32,8 +32,9 @@ import (
 	"runtime"
 
 	ctrdlog "github.com/containerd/containerd/log"
-	fccdcri "github.com/ease-lab/vhive/cri"
-	fccoor "github.com/ease-lab/vhive/cri/coordinator/firecracker"
+	cri "github.com/ease-lab/vhive/cri"
+	fccri "github.com/ease-lab/vhive/cri/firecracker"
+
 	ctriface "github.com/ease-lab/vhive/ctriface"
 	hpb "github.com/ease-lab/vhive/examples/protobuf/helloworld"
 	pb "github.com/ease-lab/vhive/proto"
@@ -153,9 +154,12 @@ func criServe() {
 
 	s := grpc.NewServer()
 
-	coor := fccoor.NewFirecrackerCoordinator(orch)
+	fcService, err := fccri.NewFirecrackerService(orch)
+	if err != nil {
+		log.Fatalf("failed to create FirecrackerCRI service %v", err)
+	}
 
-	criService, err := fccdcri.NewService(coor)
+	criService, err := cri.NewService(fcService)
 	if err != nil {
 		log.Fatalf("failed to create CRI service %v", err)
 	}
